+++ gaitmap_datasets/utils/__init__.py
@@ -1,5 +1,5 @@
 """Common utils shared by all datasets."""
 
-from gaitmap_datasets.utils.event_detection import convert_sampling_rates_event_list
+from gaitmap_datasets.utils.event_detection import convert_segmented_stride_list
 
-__all__ = ["convert_sampling_rates_event_list"]
+__all__ = ["convert_segmented_stride_list"]