+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) (+ the Migration Guide),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - 
+
+### Fixed
+- Fixed an issue that the index of `EgaitParameterValidation2013` was not deterministically sorted.
+
 ## [0.9.0] - 2023-02-02
 
 ## Added